requirements.txt
appdirs==1.4.3
mysql-connector-python-rf==2.2.2
packaging==16.8
pymongo==3.4.0
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
six==1.10.0
